requirements.txt
Django==5.2.6
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
channels==4.1.0
channels-redis==4.2.0
redis==5.0.1
celery==5.3.6
django-celery-results==2.5.1
django-celery-beat==2.5.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
django-filter==23.5
django-tables2==2.7.0
django-import-export==4.0.1
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9
python-dateutil==2.8.2
pytz==2023.3
python-dotenv==1.0.0

# Testing
pytest==8.0.0
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-factoryboy==2.5.1
factory-boy==3.3.0
Faker==19.9.0
coverage==7.3.2

# Code quality
black==23.9.1
isort==5.12.0
flake8==6.1.0
pre-commit==3.3.3
sentry-sdk==1.32.0
django-silk==5.0.0
django-allauth==0.58.2
dj-rest-auth==4.0.1
sendgrid==6.11.0
django-sendgrid-v5==0.10.0
python-slugify==8.0.1
django-widget-tweaks==1.5.0
django-model-utils==4.3.1
django-extensions==3.2.3
django-csp==5.0.0
django-axes==6.0.2
django-ratelimit==4.0.0
django-celery-beat==2.8.1